[Error] Update value of x to a new floating point number.
Previously I updated the variable to an int from a floating point number which gave me an error. Now I updated it to a floating point number, but not I got an error saying that it cannot be initialized more than once. This is immutability by default in Rust to make code memory safe.